requirements.txt
beautifulsoup4==4.11.1
binance==0.3
python_binance==1.0.16
python_telegram_bot==13.13
requests==2.28.1
tradingview_ta==3.2.10
lunarcrush==1.0.3
pandas==1.4.3
ta-lib==0.4.24